% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client}
\alias{client}
\title{Create a client connection to the Segment.IO API}
\usage{
client(url = "https://api.segment.io",
  writekey = Sys.getenv("SEGMENT_WRITEKEY"), debugging = FALSE)
}
\arguments{
\item{url}{The URL to the Segment.IO API service}

\item{writekey}{the writekey for authorization to post to Segment.IO}

\item{debugging}{set to TRUE to increase stdout logging and provide additional debugging detail}
}
\description{
Create a client connection to the Segment.IO API
}
\examples{
client(
  writekey = 'fake_key'
)
}
