% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify}
\alias{identify}
\title{Create a client connection to the Sittercity API}
\usage{
identify(userId, traits = list(), cli = client())
}
\arguments{
\item{userId}{a unique string to identify user who should be identified}

\item{traits}{a list of items to write as traits of the provided user}

\item{cli}{the Segment.IO client [created from ::client]}
}
\description{
Create a client connection to the Sittercity API
}
\examples{
identify(
  userId = 'fake_uuid',
  traits = list(status = 'closed')
)
}
