% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{SegmentIOClient-class}
\alias{SegmentIOClient-class}
\title{An S4 class to represent a client connection}
\description{
An S4 class to represent a client connection
}
\section{Slots}{

\describe{
\item{\code{authorization}}{the Base64 encoded authorization string}

\item{\code{base_url}}{the url for c}
}}

